refactor: move utils to services for use throughout API
