+++ services/util.py
@@ -1,0 +1,98 @@
+from shapely.ops import transform
+import pyproj
+import httpx
+
+TRANSFORMERS = {}
+
+
+def transform_srid(geometry, source_srid, target_srid):
+    """
+    geometry must be a shapely geometry object, like Point, Polygon, or MultiPolygon
+    """
+    transformer_key = (source_srid, target_srid)
+    if transformer_key not in TRANSFORMERS:
+        source_crs = pyproj.CRS(f"EPSG:{source_srid}")
+        target_crs = pyproj.CRS(f"EPSG:{target_srid}")
+        transformer = pyproj.Transformer.from_crs(
+            source_crs, target_crs, always_xy=True
+        )
+        TRANSFORMERS[transformer_key] = transformer
+    else:
+        transformer = TRANSFORMERS[transformer_key]
+    return transform(transformer.transform, geometry)
+
+
+def get_tiger_data(
+    lon: float, lat: float, layer: int, outfields: str = "*"
+) -> dict | None:
+    url = f"https://tigerweb.geo.census.gov/arcgis/rest/services/TIGERweb/State_County/MapServer/{layer}/query"
+    params = {
+        "f": "json",
+        "where": "1=1",
+        "geometry": f"{lon},{lat}",
+        "geometryType": "esriGeometryPoint",
+        "inSR": "4326",
+        "spatialRel": "esriSpatialRelIntersects",
+        "outFields": outfields,
+        "returnGeometry": "false",
+    }
+    resp = httpx.get(url, params=params, timeout=15)
+    data = resp.json()
+    if not data.get("features"):
+        return None
+
+    return data["features"][0]["attributes"]
+
+
+def get_state_from_point(lon: float, lat: float) -> str:
+    attrs = get_tiger_data(lon, lat, layer=0, outfields="BASENAME")
+    return attrs["BASENAME"]
+
+
+def get_county_from_point(lon: float, lat: float) -> str:
+    """
+    Look up county for a given longitude/latitude
+    using the US Census TIGERWeb REST API.
+    """
+
+    attrs = get_tiger_data(lon, lat, layer=1, outfields="BASENAME")
+    return attrs["BASENAME"]
+
+
+def get_quad_name_from_point(lon: float, lat: float) -> str:
+    url = "https://carto.nationalmap.gov/arcgis/rest/services/map_indices/MapServer/10/query"
+    params = {
+        "f": "json",
+        "geometry": f"{lon},{lat}",
+        "geometryType": "esriGeometryPoint",
+        "inSR": "4326",
+        "spatialRel": "esriSpatialRelIntersects",
+        "outFields": "CELL_NAME,CELL_MAPCODE",
+        "returnGeometry": "false",
+    }
+
+    resp = httpx.get(url, params=params, timeout=15)
+    data = resp.json()
+
+    if data["features"]:
+        attrs = data["features"][0]["attributes"]
+        return attrs["CELL_NAME"]
+    else:
+        print(f"No quad name found for POINT ({lon} {lat})")
+        return None
+
+
+def get_epqs_elevation(lon: float, lat: float) -> float:
+    url = "https://epqs.nationalmap.gov/v1/json"
+    params = {
+        "x": lon,
+        "y": lat,
+        "units": "Meters",
+        "wkid": "4326",
+        "includeDate": False,
+    }
+
+    resp = httpx.get(url, params=params)
+    data = resp.json()
+
+    return data["value"]
