+++ transfers/util.py
@@ -17,15 +17,14 @@ from datetime import datetime
 import re
 from pathlib import Path
 import logging
-import httpx
-import pyproj
 from shapely import Point
-from shapely.ops import transform
+
 
 from sqlalchemy.orm import Session
 import pandas as pd
 
 from db import Thing, Location
+from services.util import transform_srid, get_epqs_elevation
 
 log_filename = f"transfers/transfer_{datetime.now():%Y-%m-%dT%Hh%Mm%Ss}.log"
 
@@ -40,29 +39,10 @@ logging.basicConfig(
 )
 logger = logging.getLogger(__name__)
 
-TRANSFORMERS = {}
-
 
 def read_csv(name: str) -> pd.DataFrame:
     p = Path(".") / "transfers" / "data" / name
     return pd.read_csv(p)
-
-
-def transform_srid(geometry, source_srid, target_srid):
-    """
-    geometry must be a shapely geometry object, like Point, Polygon, or MultiPolygon
-    """
-    transformer_key = (source_srid, target_srid)
-    if transformer_key not in TRANSFORMERS:
-        source_crs = pyproj.CRS(f"EPSG:{source_srid}")
-        target_crs = pyproj.CRS(f"EPSG:{target_srid}")
-        transformer = pyproj.Transformer.from_crs(
-            source_crs, target_crs, always_xy=True
-        )
-        TRANSFORMERS[transformer_key] = transformer
-    else:
-        transformer = TRANSFORMERS[transformer_key]
-    return transform(transformer.transform, geometry)
 
 
 def get_valid_point_ids(session, thing_type="water well"):
@@ -103,82 +83,6 @@ def convert_to_wgs84_vertical_datum(row, z):
     elif row.VerticalDatum == "NGVD29":
         z = z + 3.0  # TODO: check this transformation
     return z
-
-
-def get_state_from_point(lon: float, lat: float) -> str:
-    attrs = get_tiger_data(lon, lat, layer=0, outfields="BASENAME")
-    return attrs["BASENAME"]
-
-
-def get_county_from_point(lon: float, lat: float) -> str:
-    """
-    Look up county for a given longitude/latitude
-    using the US Census TIGERWeb REST API.
-    """
-
-    attrs = get_tiger_data(lon, lat, layer=1, outfields="BASENAME")
-    return attrs["BASENAME"]
-
-
-def get_tiger_data(
-    lon: float, lat: float, layer: int, outfields: str = "*"
-) -> dict | None:
-    url = f"https://tigerweb.geo.census.gov/arcgis/rest/services/TIGERweb/State_County/MapServer/{layer}/query"
-    params = {
-        "f": "json",
-        "where": "1=1",
-        "geometry": f"{lon},{lat}",
-        "geometryType": "esriGeometryPoint",
-        "inSR": "4326",
-        "spatialRel": "esriSpatialRelIntersects",
-        "outFields": outfields,
-        "returnGeometry": "false",
-    }
-    resp = httpx.get(url, params=params, timeout=15)
-    data = resp.json()
-    if not data.get("features"):
-        return None
-
-    return data["features"][0]["attributes"]
-
-
-def get_quad_name_from_point(lon: float, lat: float) -> str:
-    url = "https://carto.nationalmap.gov/arcgis/rest/services/map_indices/MapServer/10/query"
-    params = {
-        "f": "json",
-        "geometry": f"{lon},{lat}",
-        "geometryType": "esriGeometryPoint",
-        "inSR": "4326",
-        "spatialRel": "esriSpatialRelIntersects",
-        "outFields": "CELL_NAME,CELL_MAPCODE",
-        "returnGeometry": "false",
-    }
-
-    resp = httpx.get(url, params=params, timeout=15)
-    logger.info(resp)
-    data = resp.json()
-
-    if data["features"]:
-        attrs = data["features"][0]["attributes"]
-        return attrs["CELL_NAME"]
-    else:
-        logger.warning(f"No quad name found for POINT ({lon} {lat})")
-
-
-def get_epqs_elevation(lon: float, lat: float) -> float:
-    url = "https://epqs.nationalmap.gov/v1/json"
-    params = {
-        "x": lon,
-        "y": lat,
-        "units": "Meters",
-        "wkid": "4326",
-        "includeDate": False,
-    }
-
-    resp = httpx.get(url, params=params)
-    data = resp.json()
-
-    return data["value"]
 
 
 def make_location(row: pd.Series) -> Location:
